+++ package.json
@@ -21,7 +21,6 @@
     "react-router-dom": "^6.5.0",
     "react-scripts": "5.0.1",
     "recharts": "^2.2.0",
-    "styled-components": "^5.3.6",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
