--- src/lib/scale.js
@@ -1,5 +1,0 @@
-export const s = scale;
-export const vs = verticalScale;
-
-export const getScaledWidthPixel = (size) => `${(size)}px`;
-export const getScaledHeightPixel = (size) => `${(size)}px`;
