+++ src/pages/node/components/StickyHeader.js
@@ -8,7 +8,7 @@ import TableHead from "@mui/material/TableHead";
 import TablePagination from "@mui/material/TablePagination";
 import TableRow from "@mui/material/TableRow";
 
-import { nodeMockData } from "./utils/nodeMockData";
+import { nodeMockData } from "../utils/nodeMockData";
 
 const columns = [
   {
