--- src/pages/node/utils/nodeTable.js
@@ -1,47 +1,0 @@
-import React, { useState } from "react";
-
-import Box from "@mui/material/Box";
-import Table from "@mui/material/Table";
-import TableContainer from "@mui/material/TableContainer";
-import Paper from "@mui/material/Paper";
-
-import StickyHeadTable from "../StickyHeader";
-import { nodeMockData } from "./nodeMockData";
-
-const NodeTable = () => {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(5);
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
-
-  const emptyRows =
-    page > 0
-      ? Math.max(0, (1 + page) * rowsPerPage - nodeMockData.nodeList.length)
-      : 0;
-
-  return (
-    <Box sx={{ width: "90%" }}>
-      <Paper sx={{ width: "100%", mb: 2 }}>
-        <TableContainer>
-          <Table
-            sx={{ minWidth: 750 }}
-            aria-labelledby="tableTitle"
-            size={"medium"}
-          >
-            <StickyHeadTable/>
-            
-          </Table>
-        </TableContainer>
-      </Paper>
-    </Box>
-  );
-};
-
-export default NodeTable;
